api/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
httpx
wikipedia
beautifulsoup4
